Support regr_r2 aggregate function
