+++ docs/velox-backend-support-progress.md
@@ -353,7 +353,7 @@ Gluten supports 199 functions. (Draw to right to see all data types)
 | bool_or                       |                        |                       |        |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
 | collect_list                  |                        |                       | S      |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
 | collect_set                   |                        |                       | S      |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
-| corr                          | corr                   |                       | S      |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
+| corr                          | corr                   |                       | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
 | count                         | count                  |                       | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
 | count_if                      | count_if               |                       |        |                        |         | S    | S     | S   | S    | S     |        |      |           |        |         |      |        |          |       |      |        |      |
 | count_min_sketch              |                        |                       |        |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
@@ -372,6 +372,7 @@ Gluten supports 199 functions. (Draw to right to see all data types)
 | mean                          | avg                    |                       | S      | Ansi Off               |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
 | min                           | min                    |                       | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
 | min_by                        |                        |                       | S      |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
+| regr_r2                       | regr_r2                | regr_r2               | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
 | skewness                      | skewness               | skewness              | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
 | some                          |                        |                       |        |                        |         |      |       |     |      |       |        |      |           |        |         |      |        |          |       |      |        |      |
 | std,stddev                    | stddev                 |                       | S      |                        |         |      | S     | S   | S    | S     | S      |      |           |        |         |      |        |          |       |      |        |      |
